requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.9.0
sqlalchemy>=1.4.0
alembic>=1.7.0
psycopg2-binary>=2.9.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.5
celery>=5.2.0
redis>=4.0.0
numpy>=1.21.0
scipy>=1.7.0
torch>=1.9.0
minio>=7.1.0
python-dotenv>=0.19.0
pytest>=6.2.0
h5py>=3.7.0
matplotlib>=3.5.0
pytest-asyncio>=0.18.0
httpx>=0.23.0 